hotfix层订阅mono层消息也改成action回调方式

+++ Unity/Assets/Scripts/Event/NumericChangeEvent_NotifyWatcher.cs
@@ -1,7 +1,7 @@
 ﻿namespace Model
 {
 	// 分发数值监听
-	[Event((int)EventIdType.NumbericChange)]
+	[Event(EventIdType.NumbericChange)]
 	public class NumericChangeEvent_NotifyWatcher: AEvent<NumericType, long, int>
 	{
 		public override void Run(NumericType numericType, long id, int value)
